+++ tests/test_interface.py
@@ -6,7 +6,6 @@ import numpy as np
 
 @pytest.mark.parametrize("optimize_type", ["preset", "list", "tuple"])
 def test_array_contract_path_cache(optimize_type):
-
     if optimize_type == "preset":
         optimize = "auto"
     elif optimize_type == "list":
@@ -35,7 +34,6 @@ def test_array_contract_path_cache(optimize_type):
 
 @pytest.mark.parametrize("optimize_type", ["preset", "list", "tuple"])
 def test_array_contract_expression_cache(optimize_type):
-
     if optimize_type == "preset":
         optimize = "auto"
     elif optimize_type == "list":
@@ -46,13 +44,25 @@ def test_array_contract_expression_cache(optimize_type):
     inputs, output, shapes, size_dict = ctg.utils.rand_equation(10, 3)
     arrays = ctg.utils.make_arrays_from_inputs(inputs, size_dict)
     expra = ctg.array_contract_expression(
-        inputs, output, shapes=shapes, cache=True, optimize=optimize,
+        inputs,
+        output,
+        shapes=shapes,
+        cache=True,
+        optimize=optimize,
     )
     exprb = ctg.array_contract_expression(
-        inputs, output, shapes=shapes, cache=True, optimize=optimize,
+        inputs,
+        output,
+        shapes=shapes,
+        cache=True,
+        optimize=optimize,
     )
     exprc = ctg.array_contract_expression(
-        inputs, output, shapes=shapes, cache=False, optimize=optimize,
+        inputs,
+        output,
+        shapes=shapes,
+        cache=False,
+        optimize=optimize,
     )
     assert expra is exprb
     assert exprb is not exprc
@@ -62,3 +72,32 @@ def test_array_contract_expression_cache(optimize_type):
     assert np.allclose(xa, xb)
     xc = expra(*arrays)
     assert np.allclose(xa, xc)
+
+
+def test_einsum_formats_interleaved():
+    args = (
+        np.random.rand(2, 3, 4),
+        [2, 3, 4],
+        np.random.rand(4, 5, 6),
+        [4, 5, 6],
+        np.random.rand(2, 7),
+        [2, 7],
+        [7, 3],
+    )
+    x = np.einsum(*args)
+    y = ctg.einsum(*args)
+    assert np.allclose(x, y)
+
+
+@pytest.mark.parametrize("eq,shapes", [
+    ("c...a,b...c->b...a", [(2, 5, 6, 3), (4, 6, 2)]),
+    ("a...a->...", [(3, 3)]),
+    ("a...a->...a", [(3, 4, 5, 3)]),
+    ("...,...ab->ba...", [(), (2, 3, 4, 5)]),
+    ("a,b,ab...c->b...a", [(2,), (3,), (2, 3, 4, 5, 6)]),
+])
+def test_einsum_ellipses(eq, shapes):
+    arrays = [np.random.rand(*shape) for shape in shapes]
+    x = np.einsum(eq, *arrays)
+    y = ctg.einsum(eq, *arrays)
+    assert np.allclose(x, y)